Add error handling to FetchPullRequest
If the git repo can not be initialized or the
base.ref branch can not be fetched, or the refs
for the pull request can not be fetched, then
return an error from FetchPullRequest.

If the event handlers, if FetchPullRequest
returns an error, then skip the event being
processed.  For commit_comment events, skip
the current PR and try the next PR found in
the search.

Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>